# api/public/swagger.yaml
components:
...
    schemas:
        TokenResponse:
            properties:
                accessToken:
                    type: string
                    description: 'Access Token'
                    example: someRandomCryptoString
                refreshToken:
                    type: string
                    description: 'Refresh Token'
                    example: someRandomCryptoString
            required:
                - accessToken
                - refreshToken
            type: object
            additionalProperties: false
        TokenPayload:
            properties:
                clientId:
                    type: string
                    description: 'Client ID'
                    example: clientID1
                code:
                    type: string
                    description: 'Authorization code'
                    example: someRandomCryptoString
            required:
                - clientId
                - code
            type: object
            additionalProperties: false
...
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
info:
    title: api
    version: 0.0.2
    description: 'Api of SASjs server'
    contact:
        name: '4GL Ltd'
openapi: 3.0.0
paths:
    /SASjsApi/auth/token:
        post:
            operationId: Token
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TokenResponse'
                            examples:
                                'Example 1':
                                    value: {accessToken: someRandomCryptoString, refreshToken: someRandomCryptoString}
            summary: 'Accepts client/auth code and returns access/refresh tokens'
            tags:
                - Auth
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/TokenPayload'
    /SASjsApi/auth/refresh:
        post:
            operationId: Refresh
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TokenResponse'
                            examples:
                                'Example 1':
                                    value: {accessToken: someRandomCryptoString, refreshToken: someRandomCryptoString}
            summary: 'Returns new access/refresh tokens'
            tags:
                - Auth
            security:
                -
                    bearerAuth: []
            parameters: []
    /SASjsApi/auth/logout:
        post:
            operationId: Logout
            responses:
                '204':
                    description: 'No content'
            summary: 'Logout terminate access/refresh tokens and returns nothing'
            tags:
                - Auth
            security:
                -
                    bearerAuth: []
            parameters: []
...
servers:
    -
        url: /
...